refactor(service): 서비스 인터페이스 제거 #23
- 구현체로 바로 사용할 수 있도록 인터페이스 제거 및 기존의 Impl 클래스를 서비스 클래스로 사용
